fifth commit

+++ controllers/contactcontrollers.js
@@ -3,17 +3,17 @@ const Contact = require("../models/contactModel")
 
 // @desc Get all contacts
 // @route GET /api/contacts
-// @access public
+// @access private
 
 const getContacts = asyncHandler( async (req,res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 });
 // ------------------------------
 
 // @desc Create new contacts
 // @route POST /api/contacts
-// @access public
+// @access private
 
 const createContact = asyncHandler(async (req,res) => {
     const {name,email,phone} = req.body;
@@ -23,6 +23,7 @@ const createContact = asyncHandler(async (req,res) => {
     }
     const contact = await Contact.create({
         name,email,phone,
+        user_id:req.user.id
     });
     res.status(201).json(contact);
 });
@@ -31,7 +32,7 @@ const createContact = asyncHandler(async (req,res) => {
 
 // @desc Get contact
 // @route GET /api/contacts/:id
-// @access public
+// @access private
 
 const getContact = asyncHandler( async (req,res) => {
     const contact = await Contact.findById(req.params.id);
@@ -46,13 +47,17 @@ const getContact = asyncHandler( async (req,res) => {
 
 // @desc Update contact
 // @route PUT /api/contacts/:id
-// @access public
+// @access private
 
 const updateContact = asyncHandler( async (req,res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
+    }
+    if(contact.user_id.toString() !== req.user_id){
+        res.status(403);
+        throw new Error("User dont have permission to update other user contacts");
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -68,7 +73,7 @@ const updateContact = asyncHandler( async (req,res) => {
 
 // @desc Delete contact
 // @route DELETE /api/contacts/:id
-// @access public
+// @access private
 
 const deleteContact = asyncHandler( async (req,res) => {
     const contact = await Contact.findById(req.params.id);
@@ -76,6 +81,11 @@ const deleteContact = asyncHandler( async (req,res) => {
         res.status(404);
         throw new Error("Contact not found");
     }
+    if(contact.user_id.toString() !== req.user_id){
+        res.status(403);
+        throw new Error("User dont have permission to update other user contacts");
+    }
+    
     res.status(200).json(contact);
     await contact.deleteOne();
 });
